break up grammar definitions more
